requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
click==8.1.7
fastapi==0.115.5
graphql-core==3.2.5
h11==0.14.0
idna==3.10
libcst==1.5.1
markdown-it-py==3.0.0
mdurl==0.1.2
pydantic==2.10.1
pydantic_core==2.27.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-multipart==0.0.17
PyYAML==6.0.2
rich==13.9.4
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.41.3
strawberry-graphql==0.253.0
typer==0.13.1
typing_extensions==4.12.2
uvicorn==0.32.1
